Divide JS File
