+++ index.html
@@ -44,5 +44,5 @@
       </div>
     </div>
   </body>
-  <script type="module" src="./src/index.js" type="text/javascript"></script>
+  <script type="module" src="./src/main.js"></script>
 </html>
