+++ src/state.js
@@ -1,0 +1,7 @@
+export let state = {
+  isSearching: false,
+  currentPage: 1,
+  total_pages: 0,
+  currentSearchQuery: '',
+  movieData: [],
+};
