+++ src/tmdb-api/option.js
@@ -1,0 +1,8 @@
+export const options = {
+  method: 'GET',
+  headers: {
+    accept: 'application/json',
+    Authorization:
+      'Bearer eyJhbGciOiJIUzI1NiJ9.eyJhdWQiOiIyOWQ3YWI1OWZmZTc5YmQyODA5NTRhMjQ4M2QzOWUxZSIsInN1YiI6IjYxYWIxYjkwNmMxZTA0MDA0MmI2Yzg0NyIsInNjb3BlcyI6WyJhcGlfcmVhZCJdLCJ2ZXJzaW9uIjoxfQ._rYfZK7OtFsmskyV66wMO10vbon3g-WxZSzZC99bEtg',
+  },
+};
